Renamed utils.js to index.js
